Drop the paperclip dependency
The gem does not really depend on paperclip.
