+++ artwork.gemspec
@@ -21,8 +21,9 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'bundler', '~> 1.6'
   spec.add_development_dependency 'rake'
   spec.add_development_dependency 'rspec', '~> 3'
+  spec.add_development_dependency 'paperclip', '>= 2.3'
+  spec.add_development_dependency 'sqlite3'
 
   spec.add_dependency 'rails', '>= 2.3'
-  spec.add_dependency 'paperclip', '>= 2.3'
   spec.add_dependency 'uglifier'
 end
