+++ spec/spec_helper.rb
@@ -1,2 +1,3 @@
 require 'rspec'
 require 'artwork'
+require 'shared/model'
